Move fetchSearchResults to its own module
